PODS-6186 Move to premigration package
